+++ src/aui-base/HISTORY.md
@@ -6,6 +6,14 @@
 
 No registries yet.
 
+## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)
+
+No changes.
+
+## [3.0.2](https://github.com/liferay/alloy-ui/releases/tag/3.0.2)
+
+No changes.
+
 ## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)
 
 * [AUI-1758](https://issues.liferay.com/browse/AUI-1758) LayoutBuilder should allow all breakpoints and should not have a min width