+++ src/aui-form-validator/HISTORY.md
@@ -4,6 +4,14 @@
 
 ## @VERSION@
 
+No registries yet.
+
+## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)
+
+* [AUI-2049](https://issues.liferay.com/browse/AUI-2049) Remove custom and requireValidation validator properties
+
+## [3.0.2](https://github.com/liferay/alloy-ui/releases/tag/3.0.2)
+
 * [AUI-2027](https://issues.liferay.com/browse/AUI-2027) Allow requiring custom validation without field values
 * [AUI-2040](https://issues.liferay.com/browse/AUI-2040) Change validation for file input on change event