+++ src/aui-io/HISTORY.md
@@ -6,6 +6,15 @@
 
 No registries yet.
 
+## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)
+
+No changes.
+
+## [3.0.2](https://github.com/liferay/alloy-ui/releases/tag/3.0.2)
+
+No changes.
+
+
 ## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)
 
 No changes.