+++ src/aui-tree/HISTORY.md
@@ -6,6 +6,14 @@
 
 No registries yet.
 
+## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)
+
+No changes.
+
+## [3.0.2](https://github.com/liferay/alloy-ui/releases/tag/3.0.2)
+
+No changes.
+
 ## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)
 
 * [AUI-835](https://issues.liferay.com/browse/AUI-835) Alloy Tree can not be used without the mouse